# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for Auth Blueprint.",
        "title": "Auth Blueprint Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Roman Kasovsky",
            "email": "roman@kasovsky.ru"
        },
        "license": {
            "name": "Apache-2.0",
            "url": "https://directory.fsf.org/wiki/License:Apache-2.0"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/delete": {
            "post": {
                "description": "Delete Description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Delete Summary",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Refresh Description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh Summary",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.TokenData"
                        }
                    }
                }
            }
        },
        "/auth/signin": {
            "post": {
                "description": "Signin Description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Signin Summary",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.TokenData"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Signup Description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Signup Summary",
                "parameters": [
                    {
                        "description": "Register",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.Register"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/auth/welcome": {
            "get": {
                "description": "Welcome Description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Welcome Summary",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ClaimData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/other/healthcheck": {
            "get": {
                "description": "HealthCheck Description",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "other"
                ],
                "summary": "HealthCheck Summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.ClaimData": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "auth.Credentials": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth.Register": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth.TokenData": {
            "type": "object",
            "properties": {
                "expiration": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    }
}